+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [1.37.0] - 2023-11-06
+
 ### Added
 
 - add an authentication metric to check if the access is authenticated